add move details store to capture player moves

+++ src/main/java/com/dai/llew/kalah/store/GameStore.java
@@ -12,5 +12,5 @@ public interface GameStore {
 
     void saveGame(Game game);
 
-    Game getGameByID(long id);
+    Game getGameByID(int id);
 }
